#3 -- Fractal Directory Structure Redlight
Switch to fractal directory structure

Needed clarity and simplicity before building factory pattern

Switching to objects with default parameters and methods instead of pure functions, using 'this'
