+++ api/tests/results/validate.test.js
@@ -1,0 +1,7 @@
+const { validate } = require("../routes/results/validate");
+const { factory } = require("../routes/results/factory");
+
+tap.test("validate() returns valid types", t => {
+  validate(factory.valid.single);
+  t.end();
+});
