+++ api/utilities/frp.js
@@ -8,11 +8,11 @@ const pipe = (...fns) => x => fns.reduce((v, f) => f(v), x);
  * @param {string} key Names top level index and selects whitelist
  * @param  {Object[]} objects Merges under `key`
  */
-const nest = (key, objects) => {
+const whitelist = (key, objects) => {
   return {
     [key]: objects
       .filter(object => fields[key].includes(Object.keys(object)[0]))
       .reduce((cur, prev) => ({ ...cur, ...prev }), {}),
   };
 };
-module.exports = { pipe, nest };
+module.exports = { pipe, whitelist };
